Fixed minor error.

+++ src/questions/StreamInsteadOfFor.java
@@ -34,7 +34,7 @@ public class StreamInsteadOfFor {
     private void simpleStream2() {
         List<Integer> numbers = Arrays.asList(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
 
-        int ave = 0;
+        double ave = 0.0;
         for (Integer x : numbers) {
             ave += x;
         }
